requirements.txt
asgiref==3.8.1
Django==5.2.1
django-restframework==0.0.1
sqlparse==0.5.3
